Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -9,7 +9,7 @@ int welcome(int a, int b, char c) {
         case '*':
             return a*b;
         case '/':
-            return a/b;      
+            return (int)a/b;      
         default:
             return "error";          
     }
